% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_lift_chart}
\alias{create_lift_chart}
\title{Internal function: Create lift-chart ggplot2 object}
\usage{
create_lift_chart(score, binary, prop, xlab, ylab, title)
}
\arguments{
\item{score}{A vector inherits `score` in `ICLC`}

\item{binary}{A vector inherits `binary` in `ICLC`}

\item{prop}{A numeric inherits `prop` in `ICLC`}

\item{xlab}{A character represents plot's x axis label}

\item{ylab}{A character represents plot's y axis label}

\item{title}{A character represents plot's title}
}
\value{
A ggplot2 object
}
\description{
Internal function: Create lift-chart ggplot2 object
}
\keyword{internal}
