% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_output_indeces}
\alias{fetch_output_indeces}
\title{Internal function: Fetch indeces in an output object}
\usage{
fetch_output_indeces(obj)
}
\arguments{
\item{obj}{A valid object (e.g., `IAUC`, `LAUC`)}
}
\description{
Internal function: Fetch indeces in an output object
}
\keyword{internal}
