% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{plot.ICLC}
\alias{plot.ICLC}
\title{Visualize ICLC results}
\usage{
\method{plot}{ICLC}(x, ...)
}
\arguments{
\item{x}{An ICLC class object}

\item{...}{Not used directly}
}
\description{
Visualize ggplot2 objects in ICLC sequentially
}
\examples{
library(ROCR)
data("ROCR.simple")
Coutput <- ICLC(ROCR.simple$predictions, ROCR.simple$labels)
plot(Coutput)
}
\seealso{
\link{ICLC}
}
