% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_sequentially}
\alias{plot_sequentially}
\title{Internal function: Plot sequentially}
\usage{
plot_sequentially(objs)
}
\arguments{
\item{objs}{A valid object (e.g., `IAUC`, `LAUC`)}
}
\description{
Internal function: Plot sequentially
}
\keyword{internal}
