% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{print.IAUC}
\alias{print.IAUC}
\title{Show IAUC results}
\usage{
\method{print}{IAUC}(x, ...)
}
\arguments{
\item{x}{An IAUC class object for `print method}

\item{...}{Not used directly}
}
\description{
Print IAUC output in detail
}
\examples{
library(ROCR)
data("ROCR.simple")
Ioutput <- IAUC(ROCR.simple$predictions, ROCR.simple$labels)
print(Ioutput)
}
\seealso{
\link{IAUC}
}
