% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{print.LAUC}
\alias{print.LAUC}
\title{Show LAUC results}
\usage{
\method{print}{LAUC}(x, ...)
}
\arguments{
\item{x}{An LAUC class object for `print` method}

\item{...}{Not used directly}
}
\description{
Print LAUC output in detail
}
\examples{
library(ROCR)
data("ROCR.simple")
Loutput <- LAUC(ROCR.simple$predictions, ROCR.simple$labels)
print(Loutput)
}
\seealso{
\link{LAUC}
}
