% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_output}
\alias{print_output}
\title{Internal function: Print output}
\usage{
print_output(output, attrs)
}
\arguments{
\item{output}{A valid object (e.g., `IAUC`, `LAUC`)}

\item{attrs}{A vector of attribute names}
}
\description{
Internal function: Print output
}
\keyword{internal}
